Create ContextAPI
The Context API simplifies the process of managing shared state or configuration across components. It is especially useful when dealing with deeply nested components or when multiple components need access to the same data.

Context Creation: You create a context using the createContext() function provided by React. This creates a context object that has a Provider component and a Consumer component.

Provider Component: The Provider component is used to wrap the components that need access to the shared data. It accepts a value prop which defines the data to be shared.

Consumer Component: The Consumer component is used within the wrapped components to access the shared data from the context. It uses a render prop or the useContext hook to consume the context and retrieve the data.

Data Propagation: The context's Provider component ensures that the provided data is accessible to all descendant components that consume the context, regardless of their level in the component tree. This eliminates the need to pass props manually through intermediate components.